updated meta data on product page make the layout to server side
