+++ src/components/ui/footer/footer.tsx
@@ -1,3 +1,4 @@
+"use client";
 import React, { useEffect, useState } from "react";
 // import JumpingButtons from "../common/jumping-buttons";
 import renderImg from "@/imgImport";
